Complete bonus task (lab 2); app_elixir refactoring
