+++ app_elixir/.gitignore
@@ -20,7 +20,7 @@ erl_crash.dump
 *.ez
 
 # Ignore package tarball (built via "mix hex.build").
-devops-*.tar
+devops_web-*.tar
 
 # Temporary files, for example, from tests.
 /tmp/
