+++ app_elixir/config/prod.exs
@@ -1,11 +1,6 @@
 import Config
 
-# Note we also include the path to a cache manifest
-# containing the digested version of static files. This
-# manifest is generated by the `mix assets.deploy` task,
-# which you should run after static files are built and
-# before starting your production server.
-config :devops_web, DevopsWebWeb.Endpoint, cache_static_manifest: "priv/static/cache_manifest.json"
+config :devops_web, DevopsWebWeb.Endpoint, server: true
 
 # Configures Swoosh API Client
 config :swoosh, api_client: Swoosh.ApiClient.Finch, finch_name: DevopsWeb.Finch
