--- app_elixir/lib/devops.ex
@@ -1,18 +1,0 @@
-defmodule Devops do
-  @moduledoc """
-  Documentation for `Devops`.
-  """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> Devops.hello()
-      :world
-
-  """
-  def hello do
-    :world
-  end
-end
