--- app_elixir/test/devops_test.exs
@@ -1,8 +1,0 @@
-defmodule DevopsTest do
-  use ExUnit.Case
-  doctest Devops
-
-  test "greets the world" do
-    assert Devops.hello() == :world
-  end
-end
