+++ app_elixir/test/test_helper.exs
@@ -1,1 +1,2 @@
 ExUnit.start()
+# Ecto.Adapters.SQL.Sandbox.mode(DevopsWeb.Repo, :manual)
